Create TwoSumBruteForce.java
The approach is very straightforward. We loop through each element i in the array of nums and find if there is another value that equals the difference between target and i. (target - i)

The Time Complexity for this version is O(n^2). For each element, we try to find its complement by looping through the rest of the array which takes O(n) time. Therefore, the time complexity is O(n^2).

The Space Complexity for this version is O(1). The space required does not depend on the size of the input array, so only constant space is used.