Create Two-Pass Hash Table Version.java
The approach for this is to go through all the numbers in the array of nums, look at our target, subtract the value of the number that we are at from the target. And if that already exists in the HashMap, then we already have the number. Otherwise, we are going to store the number that we have. And hopefully, through more iterations, we will have the answer.

The purpose of using a Hash table is to reduce the lookup time from O(n) to O(1) by trading space for speed. A Hash table supports fast lookup at near-constant time.

The Time Complexity for this version is O(n). We traverse the list containing n elements exactly twice. Since the hash table reduces the lookup time to O(1). Therefore, the time complexity is O(n).

The Space Complexity for this version is O(n). The space required depends on the number of items stored in the hash table, which stores n elements.