Create TwoSumPointerArrayIsSorted.java
In this version, instead of using a hash table or a brute force approach which does not make use of the property that the input array is sorted. Instead, we use two indices or pointers. One pointed at the first element of the array and another pointed at the last element of the array. If the sum of these pointer values is equal to target, we found our solution. If it is less than the target, we increase the smaller index. If it is greater than target, we decrease the larger index by 1. Move the indices and repeat the comparison until the solution is found. 

And finally, the reason we return -1,-1 is because it won’t be reached or reached if we don’t have a solution. Otherwise, we return the solution that we found.

The Time Complexity for this version is O(n). We traverse the list containing n elements once.

The Space Complexity for this version is O(1). We only use additional space to store two indices and sum.