Create ThreeSum.java
This is a follow-up to Two Sum. The purpose of Three Sum is that you have an array of numbers and that you have to return a triple value as an array containing the three numbers that add up to the value and cannot have repeating values or duplicates.

So, the first step is that we need to sort the array. After that, we declare an output. In this case, I declared a List of Integers output to be what I return. Now, we iterate through the array. The first integer i cannot be greater than the numbers.length - 2 because that and numbers.length - 1 may represent the other two values. And in the case where i is greater than 0, break the loop. Remaining values cannot sum up to zero. If the current value is the same as before, we skip it.

Like the Two Sums version, we have two pointers, low and high. Low pointing to i + 1 and High pointing to the last index. We also set sum as the difference between 0 and the ith value in the numbers array. Now, we have a while loop (low < high). 

So while low is less than high, if the sum of low and high is equal to the sum. Then, we add those numbers including i to the output. But, if nums[low] == nums[low+1] in the case where we have a duplicate in the low pointer, we need to increment by 1. In the case where nums[high] == nums[high-1] where we have a duplicate in the high pointer, we must decrement by 1.
Next, we increment low pointer and decrement high pointer for the other values. And we need statements in the case where if nums[low] + nums[high] > sum, we must decrement high pointer. Otherwise, increment the low pointer. Finally, we return our output. 

The Time Complexity for this version is O(n^2). We sort the array which takes O(n log n) times and we traverse the array which is O(n) and the while loop is O(n). So overall complexity is O(n log n + n^2) and this is asymptotically equivalent to O(n^2).

The Space Complexity for this version is O(n) for the Linked List we are using in our approach.