Create AddTwoNums.java
So, you have two non-empty linked lists representing non-negative integers stored in reverse order. Each node contains a single digit. Add two numbers and return the sum as a linked list. 

To start, we will need to initialize a few ListNodes, one to keep track of results, another to keep track of previous value (prev), and lastly a temporary value (temp). We also need a carry as well as sum which will both be initialized to 0. 

So, while ListNodes l1 or l2 is not empty, we traverse them until both ends. If any of the lists end, in the case where one does and the other doesn’t, we then just take 0 to add to the other list element. Then, we update the carry. And now, if the sum is greater than 9, we update the sum so that it will give one's place by taking the modulo of 10 against the sum. With this, we need a new node to hold the sum. And that is where the temp is introduced. If the temp is the first node, we make that the head of the result list. If it is not the first node, then we connect that with the previous node prev. We will then set the prev for the next insertion. And now, we move both pointers to the next node. So if (l1 != null), move to the next node. The same applies with l2. And now for the last sum, if the sum > 9, there needs to be a carry so to hold that, we create a new node. Finally, we return the result.

The Time Complexity for this version is O(max(m,n)). We assume that m and n represent the length of l1 and l2 respectively, the algorithm above iterates at most max(m,n) times.

The Space Complexity for this version is O(max(m,n)) since the length of the new list is at most max(m,n) + 1.