Create SquareRoot.java
In this problem, we have an integer x and we are to return the square root of x. Since the return type is an int, it will only return the integer part of the result.

Basically, in the solution since we are not able to use any built in functions or operators. We have to use the Binary search approach. The first step is that if x  is less than or equal to 1, return x. And we need boundaries since we are using Binary Search. So our left boundary is set to 1 and our right is set to x. So while left is less than right, we set our mid or pivot (in this case mid)  as the sum of left and (right - left)/2. If mid is less than or equal to x/mid, we store mid + 1 into our left boundary. Otherwise, we store mid as our right boundary. We will then return left - 1.