Create ValidParentheses.java
In this problem, we are given a string and the only characters used are brackets. We need to find if the input string is valid if the open brackets are closed by the same brackets and in the correct order. One way to go about this problem is to use a Hash table. The reason for this is that with a hash table you can create mappings of different characters and you can call the mappings in the solution. We also need to use a stack and process each bracket of the expression one at a time. The way we process it is to iterate over a for loop where it traverses each character in the string and simply push it into the stack. This means we will process it later, and we move onto the sub-expression ahead.

Now, if we encounter a closing bracket, we check the element at top of the stack. If the element at the top of the stack is an opening bracket of the same type, we pop it off the stack and continue processing. Implying that it's a valid expression. Otherwise if the stack is empty, we set a dummy value of  #. And in the case when the mapping of the bracket does not match the top element, we will just return false. And in the case if there is another opening bracket, push to the stack. But, if the stack still contains elements, return false implying it is an invalid expression.

The Time Complexity for this version is O(n). We are traversing the given string one character at a time and push and pop operations on a stack takes O(1) time.

The Space Complexity for this version is O(n) since we are pushing all opening brackets onto the stack and in the worst case, we will end up pushing the brackets onto the stack. Ex: ((((((