Create FirstUniqueCharacterInAString.java
In this problem, we are given a string and we have to find the first non-repeating character and return its index. A way to go about solving this problem is a linear time solution. Since we are traversing the string and checking each character for uniqueness. An O(n) solution is a good solution. The idea is to go through the string and save the number of times each character appears in a Hash Map. That would take O(N) time, where N is the number of characters in a string. And then, we go through the string a second time. Using the hashmap as a reference to check if a character is unique or not. If it is, return the index. Otherwise, return -1.

The Time Complexity for this version is O(n) since we are traversing through the string of length n two times.

The Space Complexity for this version is O(1) because there is only 26 letters in the English alphabet and its constant for any length of a string.