Create RomanToInteger.java
In this problem, we are trying to convert roman numerals that are presented as characters in a string and pass it as integers. One way to solve this problem is to use a Hash Map. We store the roman numerals in this. The symbols of the roman numerals will be the key and the integer equivalents will be the value. Now, we need to initialize the variable result which we will need to return later on. Next, we use a for loop where we traverse the string. We have two cases to consider. The additive case and the subtractive case. The additive case is simple. When we traverse the string, all we do is take the character of the string and get its value from the Hash map. We add that value to the result and repeat for multiple characters. But for the subtractive case, if the character of the string is greater than the previous character of the string and if the integer i is greater than zero. We need to take the difference of the current - 2*previous and store that into the result. The reason for subtracting the previous twice from the current is that the subtractive case is just basic addition. So it looks at the previous character and adds that to the result. But to get the right answer, we need to subtract it from the result twice to get the correct answer.

The Time Complexity for this version is O(1) since there is a finite set of roman numerals. Regardless of whether or not, a for loop is present. Because the input size is finite, the time complexity is constant rather than linear.

The Space Complexity for this version is O(1) because there is only a constant number of single-value variables used.